move setup
I just tried to think through it from the new Builders perspective. You plug in your steppers and board after you assemble the machine. Then you try to move and control it.  After that you could be doing anything with it.

I know boards reads a little out of place. I think that is because they are kind of written as a pre-purchase page, but this is using them as an after assembly page. In reality I would prefer to put the nitty gritty for each board in learn (with links to that in laser and stuff), and just have the boards menu be one page with a simple chart or something, or maybe just where to plug in steppers for each board, then in learn it would be pinouts, firmware edits, setting drivers, yada yada.