Modify hardware table to indicate type of screw
I suggest modifying the hardware table to indicate type of screw (Not sure which type is best) used for the M5 x 30mm because it was unclear as to whether these should be bolts like the 14 M8's above it. I ended up purchasing bolts for these required parts, instead of screws and they are not usable in the core assembly and potentially other spots.